- Reintroduce all methods from ContentServiceImpl that were removed and moved them to a super class   to make final removal simpler.
